pyproject.toml
[project]
name = "sandbox"
version = "0.1.0"
description = "Python sandbox environment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "codeium-jupyter>=2.1.1",
    "dask[dataframe]>=2024.11.2",
    "ipywidgets>=8.1.5",
    "jupyter-ai[all]>=2.28.1",
    "jupytext>=1.16.4",
    "lightgbm>=4.5.0",
    "matplotlib>=3.9.2",
    "mlxtend>=0.23.3",
    "notebook>=7.2.2",
    "optuna>=4.1.0",
    "optunahub>=0.1.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.5.2",
    "seaborn>=0.13.2",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
]
